Assignment3/requirements.txt
pygubu==0.14
pycryptodome==3.20.0
cryptography==42.0.5
